refactor(engine): move advice-mutation helpers into a support module
