+++ crates/engine/src/exec/host.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::{BTreeMap, VecDeque},
     num::NonZeroU32,
-    sync::{Arc, Mutex},
+    sync::Arc,
 };
 
 use miden_assembly::SourceManager;
@@ -18,90 +18,10 @@ use miden_processor::{
     mast::MastForest,
 };
 
-use super::{TraceEvent, TraceHandler};
-
-/// Clone a single [AdviceMutation].
-///
-/// [AdviceMutation] does not implement [Clone] upstream, so recording and replaying event
-/// mutations requires reconstructing each variant by hand.
-pub fn clone_advice_mutation(mutation: &AdviceMutation) -> AdviceMutation {
-    match mutation {
-        AdviceMutation::ExtendStack { values } => AdviceMutation::ExtendStack {
-            values: values.clone(),
-        },
-        AdviceMutation::ExtendMap { other } => AdviceMutation::ExtendMap {
-            other: other.clone(),
-        },
-        AdviceMutation::ExtendMerkleStore { infos } => AdviceMutation::ExtendMerkleStore {
-            infos: infos.clone(),
-        },
-        AdviceMutation::ExtendPrecompileRequests { data } => {
-            AdviceMutation::ExtendPrecompileRequests { data: data.clone() }
-        }
-    }
-}
-
-/// Clone a batch of [AdviceMutation]s. See [clone_advice_mutation].
-pub fn clone_advice_mutations(mutations: &[AdviceMutation]) -> Vec<AdviceMutation> {
-    mutations.iter().map(clone_advice_mutation).collect()
-}
-
-/// A shared, cloneable log of the advice mutations produced by event handlers.
-///
-/// One entry is recorded per `on_event` invocation, in execution order, **including empty
-/// mutation sets**: event replay pops exactly one entry per event, so the log must stay aligned
-/// with the event stream of the recorded execution.
-///
-/// Attach a recorder to a live execution (see [DebuggerHost::set_event_recorder] or
-/// `DapExecutor::record_event_mutations`), run the program to completion, and feed the recorded
-/// log into [DebuggerHost::set_event_replay] (or `Executor::into_debug_with_replay`) to debug
-/// the same execution later without access to the original host's event handlers.
-#[derive(Clone, Default)]
-pub struct EventMutationRecorder {
-    log: Arc<Mutex<Vec<Vec<AdviceMutation>>>>,
-}
-
-impl EventMutationRecorder {
-    /// Create a new, empty recorder.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Returns the recorded mutation batches, leaving the recorder empty.
-    pub fn take(&self) -> Vec<Vec<AdviceMutation>> {
-        core::mem::take(&mut *self.log.lock().expect("event mutation log poisoned"))
-    }
-
-    /// Returns a copy of the recorded mutation batches, leaving the recorder intact.
-    pub fn snapshot(&self) -> Vec<Vec<AdviceMutation>> {
-        self.log
-            .lock()
-            .expect("event mutation log poisoned")
-            .iter()
-            .map(|batch| clone_advice_mutations(batch))
-            .collect()
-    }
-
-    /// The number of `on_event` invocations recorded so far.
-    pub fn len(&self) -> usize {
-        self.log.lock().expect("event mutation log poisoned").len()
-    }
-
-    /// Returns true if no `on_event` invocations have been recorded.
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Record the mutations produced by one `on_event` invocation.
-    pub(crate) fn record(&self, mutations: Vec<AdviceMutation>) {
-        self.log.lock().expect("event mutation log poisoned").push(mutations);
-    }
-
-    /// Discard everything recorded so far, e.g. when execution restarts from the beginning.
-    pub(crate) fn clear(&self) {
-        self.log.lock().expect("event mutation log poisoned").clear();
-    }
-}
+use super::{
+    TraceEvent, TraceHandler,
+    advice::{EventMutationRecorder, clone_advice_mutations},
+};
 
 /// This is an implementation of [Host] which is essentially [miden_processor::DefaultHost],
 /// but extended with additional functionality for debugging, in particular it manages trace
